# ops/surfaces/openapi.json
{
  "components": {
    "schemas": {
...
      "BrainstormItem": {
        "properties": {
          "id": {
            "description": "Identifier of the item, unique within the list. Any string.",
            "title": "Id",
            "type": "string"
          },
          "text": {
            "description": "One idea, or one organized step, as free text.",
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text"
        ],
        "title": "BrainstormItem",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectCreate": {
        "properties": {
          "name": {
            "description": "Project name, spelled exactly as tasks spell it; dots nest subprojects (`home.garden`). Surrounding whitespace is removed.",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectCreate",
        "type": "object"
      },
      "ProjectPlan": {
        "properties": {
          "brainstorm": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BrainstormItem"
            },
            "title": "Brainstorm",
            "type": "array"
          },
          "organized": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BrainstormItem"
            },
            "title": "Organized",
            "type": "array"
          },
          "principles": {
            "default": "",
            "title": "Principles",
            "type": "string"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "purpose": {
            "default": "",
            "title": "Purpose",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "vision": {
            "default": "",
            "title": "Vision",
            "type": "string"
          }
        },
        "required": [
          "project_name"
        ],
        "title": "ProjectPlan",
        "type": "object"
      },
      "ProjectPlanUpdate": {
        "properties": {
          "brainstorm": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BrainstormItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Unsorted ideas, in no particular order. The complete list: what you send replaces the stored one. Omitted or null keeps it.",
            "title": "Brainstorm"
          },
          "organized": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/BrainstormItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "The ideas worth keeping, in the order they will be acted on. The complete list: what you send replaces the stored one. Omitted or null keeps it.",
            "title": "Organized"
          },
          "principles": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The standards and constraints the work has to hold to. Omitted or null keeps the stored text; an empty string clears it.",
            "title": "Principles"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why the project exists \u2014 the outcome it serves. Omitted or null keeps the stored text; an empty string clears it.",
            "title": "Purpose"
          },
          "vision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What success looks like, described as if it had already happened. Omitted or null keeps the stored text; an empty string clears it.",
            "title": "Vision"
          }
        },
        "title": "ProjectPlanUpdate",
        "type": "object"
      },
...
      "Task": {
        "properties": {
          "annotations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TaskAnnotation"
            },
            "title": "Annotations",
            "type": "array"
          },
          "depends": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Depends",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "due": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due"
          },
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "entry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "modified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "recur": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recur"
          },
          "scheduled": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "until": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          },
          "urgency": {
            "default": 0.0,
            "title": "Urgency",
            "type": "number"
          },
          "uuid": {
            "title": "Uuid",
            "type": "string"
          },
          "wait": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wait"
          }
        },
        "required": [
          "uuid",
          "id",
          "description",
          "status"
        ],
        "title": "Task",
        "type": "object"
      },
      "TaskAnnotation": {
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "entry": {
            "title": "Entry",
            "type": "string"
          }
        },
        "required": [
          "entry",
          "description"
        ],
        "title": "TaskAnnotation",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "A self-hosted GTD task manager. Authenticate with JWT (Authorization: Bearer <token>) or API key (X-Api-Key: <key>).",
    "title": "Runway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/gtd/projects": {
      "get": {
        "description": "Return all project names \u2014 both those inferred from tasks and those created explicitly.",
        "operationId": "projects_gtd_projects_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Projects Gtd Projects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List projects",
        "tags": [
          "gtd"
        ]
      }
    },
    "/gtd/projects/{name}": {
      "get": {
        "description": "List all pending tasks belonging to a specific project.",
        "operationId": "project_tasks_gtd_projects__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 10,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tasks carrying ALL of these tags, written without `+`. Repeat the parameter for more than one. The inbox is untagged by definition, so any tag yields an empty inbox.",
              "title": "Tag"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  },
                  "title": "Response Project Tasks Gtd Projects  Name  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Project tasks",
        "tags": [
          "gtd"
        ]
      }
    },
...
    "/projects": {
      "post": {
        "description": "Create a project explicitly (it also exists implicitly as soon as a task names it). Creating the same name twice changes nothing and returns the stored plan.",
        "operationId": "create_project_projects_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a project",
        "tags": [
          "projects"
        ]
      }
    },
    "/projects/plans/{name}": {
      "get": {
        "description": "Read a project's plan (GTD Natural Planning Model: purpose, principles, vision, brainstorm, organized). Unknown names return an empty plan.",
        "operationId": "get_plan_projects_plans__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get a project plan",
        "tags": [
          "projects"
        ]
      },
      "put": {
        "description": "Create or update the plan; omitted fields are kept. It also creates the project when it does not exist yet.",
        "operationId": "upsert_plan_projects_plans__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPlanUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectPlan"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update a project plan",
        "tags": [
          "projects"
        ]
      }
    },
...
  }
}